Fork with new number formatter
A new number formatter is implemented to show custom sign symbol before
positive and negative numbers. The formatter is needed to show specific sign
symbol for negative numbers in Japanese financial report.
As the code review and merging schedule for previously submitted pull
requests regarding our new formatters are unknown, we have to fork Carbone and
maintain it as a private module.